normalized and drawn the PCA data
OK so i messed up a bit in my initial algorithm but finally tracked it
down.  There is this weird behavior the the algorithm exhibits and is
currently being patched with a shotty if statement.  Not sure if i will
go back to it.  Last step is to rotate the data over the mean